fixed
Traceback (most recent call last):
  File "adapt.py", line 74, in <module>
    main()
  File "adapt.py", line 71, in main
    s.play(player_start,player_main)
  File "adapt.py", line 30, in play
    sp.play(0,0)
  File "/home/pi/Documents/muziekinstrument/aplayer.py", line 29, in play
    old_proc.kill()
AttributeError: 'NoneType' object has no attribute 'kill'

+++ aplayer.py
@@ -26,7 +26,8 @@ class Aplayer():
         old_proc = self.processes[process_id]
         self.processes[process_id] = sp.Popen((tuple(self.get_args(pitch))),
                                               shell=False, stdout=sp.PIPE, stderr=sp.PIPE, stdin=sp.PIPE)
-        old_proc.kill()
+        if old_proc is not None:
+            old_proc.kill()
     def get_args(self,pitch):
         yield "sox"
         yield self.filename
